feat: move category related functions
